+++ src/error/internal.ts
@@ -1,0 +1,21 @@
+// tslint:disable-next-line:no-require-imports
+import setPrototypeOf = require('setprototypeof');
+import { ErrorCode } from '../errorCode';
+import { ChevreError } from './chevre';
+
+/**
+ * InternalError
+ */
+export class InternalError extends ChevreError {
+    constructor(message?: string) {
+        let actualMessage = message;
+        if (message === undefined || message.length === 0) {
+            actualMessage = 'Internal server error.';
+        }
+
+        // tslint:disable-next-line:no-single-line-block-comment
+        /* istanbul ignore next */
+        super(ErrorCode.Internal, actualMessage);
+        setPrototypeOf(this, InternalError.prototype);
+    }
+}