+++ src/errors.spec.ts
@@ -11,11 +11,13 @@ import {
     Chevre,
     Forbidden,
     GatewayTimeout,
+    Internal,
     NotFound,
     NotImplemented,
     RateLimitExceeded,
     ServiceUnavailable,
-    Unauthorized
+    Unauthorized,
+    Unknown
 } from './errors';
 
 describe('new AlreadyInUseError()', () => {
@@ -241,6 +243,44 @@ describe('new UnauthorizedError()', () => {
     it('メッセージを指定しなくても、正しくインスタンス化できる', () => {
         const error = new Unauthorized();
         assert(error instanceof Unauthorized);
+        assert.equal(error.name, 'ChevreError');
+        assert.equal(typeof error.message, 'string');
+        assert.equal(typeof error.stack, 'string');
+    });
+});
+
+describe('new InternalError()', () => {
+    it('正しくインスタンス化できる', () => {
+        const message = 'test message';
+        const error = new Internal(message);
+        assert(error instanceof Internal);
+        assert.equal(error.message, message);
+        assert.equal(error.name, 'ChevreError');
+        assert.equal(typeof error.stack, 'string');
+    });
+
+    it('メッセージを指定しなくても、正しくインスタンス化できる', () => {
+        const error = new Internal();
+        assert(error instanceof Internal);
+        assert.equal(error.name, 'ChevreError');
+        assert.equal(typeof error.message, 'string');
+        assert.equal(typeof error.stack, 'string');
+    });
+});
+
+describe('new UnknownError()', () => {
+    it('正しくインスタンス化できる', () => {
+        const message = 'test message';
+        const error = new Unknown(message);
+        assert(error instanceof Unknown);
+        assert.equal(error.message, message);
+        assert.equal(error.name, 'ChevreError');
+        assert.equal(typeof error.stack, 'string');
+    });
+
+    it('メッセージを指定しなくても、正しくインスタンス化できる', () => {
+        const error = new Unknown();
+        assert(error instanceof Unknown);
         assert.equal(error.name, 'ChevreError');
         assert.equal(typeof error.message, 'string');
         assert.equal(typeof error.stack, 'string');